refactor: Import production helpers in PR triage tests
Extract pure helper functions from test files into reusable library modules:
- scripts/automation/lib/pr-triage-helpers.js (parseConfig, extractIssuesFromBody, determineTriage, buildMetadata, generateSummary)
- scripts/automation/lib/sync-labels-helpers.js (extractPRs, validatePR, determineLabelAction, buildSyncConfig, processIssue, generateReport)
- scripts/automation/lib/integration-workflow-helpers.js (triageAndExtractIssues, syncLabelsBasedOnIssues, generateTriageSummary)

Updates to test files:
- pr-triage-orchestrator.test.js: Import from lib/pr-triage-helpers.js
- sync-pr-labels.test.js: Import from lib/sync-labels-helpers.js
- integration-workflow-pr-triage.test.js: Import from lib/integration-workflow-helpers.js

Benefits:
- Eliminates code duplication between test and production modules
- Shared implementations enable consistent testing of production behavior
- Foundation for mocking Octokit client at test boundaries
- Fixes security concern: regressions in shipped scripts can now be detected

All 104 tests passing across three test suites.
