+++ scripts/automation/__tests__/pr-triage-orchestrator.test.js
@@ -1,69 +1,15 @@
 /**
- * Pure function implementations for pr-triage-orchestrator testing
- * These functions are isolated from the production module to avoid ES module complexity
- * with Jest, but test against the same logic patterns used in production.
+ * PR Triage Orchestrator Tests
+ * Tests import production helper functions instead of duplicating them
  */
 
-function parseConfig(args) {
-  const config = {
-    dryRun: args.includes("--dry-run"),
-    verbose: args.includes("--verbose"),
-    format: "json",
-    output: ".github/reports",
-  };
-
-  const formatIdx = args.indexOf("--format");
-  if (formatIdx !== -1 && formatIdx + 1 < args.length) {
-    config.format = args[formatIdx + 1];
-  }
-
-  const outputIdx = args.indexOf("--output");
-  if (outputIdx !== -1 && outputIdx + 1 < args.length) {
-    config.output = args[outputIdx + 1];
-  }
-
-  return config;
-}
-
-function extractIssuesFromBody(body) {
-  const issues = [];
-  const regex = /(?:Fixes|Closes|Resolves|Relates to|#)[\s#]+(\d+)/gi;
-  let match;
-  while ((match = regex.exec(body)) !== null) {
-    issues.push(parseInt(match[1]));
-  }
-  return [...new Set(issues)];
-}
-
-function determineTriage(labels) {
-  return {
-    needsReview: labels.some((l) => l.name === "status:needs-review"),
-    needsChangelog: labels.some((l) => l.name === "meta:needs-changelog"),
-    isBot: labels.some((l) => l.name === "type:bot"),
-  };
-}
-
-function buildMetadata(pr, triageStatus) {
-  return {
-    prNumber: pr.number,
-    title: pr.title,
-    author: pr.user?.login,
-    needsReview: triageStatus.needsReview,
-    needsChangelog: triageStatus.needsChangelog,
-    linkedIssues: extractIssuesFromBody(pr.body || ""),
-    suggestedMilestone: triageStatus.needsReview ? "In Review" : "Backlog",
-  };
-}
-
-function generateSummary(prs, metadata) {
-  return {
-    totalPRs: prs.length,
-    reviewed: metadata.filter((m) => !m.needsReview).length,
-    needsReview: metadata.filter((m) => m.needsReview).length,
-    changelogs: metadata.filter((m) => m.needsChangelog).length,
-    linkedIssues: metadata.reduce((sum, m) => sum + m.linkedIssues.length, 0),
-  };
-}
+import {
+  parseConfig,
+  extractIssuesFromBody,
+  determineTriage,
+  buildMetadata,
+  generateSummary,
+} from "../lib/pr-triage-helpers.js";
 
 describe("pr-triage-orchestrator", () => {
   describe("parseConfig", () => {
